The main function for the not missing at random non-response

Here is the part where the description will go

Usage

mnar(
  response,
  outcome,
  calibration,
  target,
  data,
  svydesign,
  dweights,
  N,
  pop_totals,
  method = c("gencalib", "emplik", "gmm"),
  tol = 1e-08,
  maxit = 100,
  eps = .Machine$double.eps,
  control = NULL,
  ...
)

Arguments

response

response variables (instruments)

outcome

outcome variables (for modelling E(Y|X))

calibration

calibration variables (for which totals should be reproduced)

target

target variable

data

data.frame

svydesign

surveydesign object

dweights

design weights

N

population size

pop_totals

population totals (for calibration)

method

methods c("gencalib", "emplik", "gmm")

tol

tolerance 1e-8

maxit

maxit 50

eps

eps for inverse .Machine$double.eps

control

control for methods

...

TBA

Value

Returns a list with containing:

g – g-weight that sums up to sample size,

Xs – matrix used for calibration (i.e. Intercept, X and X_q transformed for calibration of quantiles),

totals – a vector of totals (i.e. N, pop_totals and pop_quantiles),

method – selected method,

backend – selected backend.

Details

Imports for the function

References

Deville, J. C., and Särndal, C. E. (1992). Calibration estimators in survey sampling. Journal of the American statistical Association, 87(418), 376-382.

add others

See also

Author

Maciej Ostapiuk, Maciej Beręsewicz

Examples

set.seed(123)
n <- 1000