Generalized calibration function based on sampling::gencalib

Importing

Usage

gencal(Xs, Zs, d, pop_totals, method = "raking", eps, maxit, tol)

Arguments

Xs

A matrix of calibration variables

Zs

A matrix of instrumental variables

d

A vector of initial weights, derived from sampling design

method

a method upon which the sampling::gencalib() function will be called

eps

describes desired accuracy level in any convergence-related

maxit

maximum number of iterations for sampling::gencalib() since it works iteratively

tol

desired tolerance

totals

A vector of population totals (including totals for both Xs and Zs)

References

Andridge, Rebecca R, and Roderick JA Little. 2011. “Proxy Pattern-Mixture Analysis for Survey Nonresponse.” Journal of Official Statistics 27 (2): 153.

Chang, T., and P. S. Kott. 2008. “Using Calibration Weighting to Adjust for Nonresponse Under a Plausible Model.” Biometrika 95 (3): 555–71. https://doi.org/10.1093/biomet/asn022.

Deville, Jean-Claude, and Carl-Erik Särndal. 1992. “Calibration Estimators in Survey Sampling.” Journal of the American Statistical Association 87 (418): 376–82.

Estevao, Victor, and Carl Särndal. n.d. “A Functional Form Approach to Calibration.” Journal of Of®cial Statistics 16 (4): 379±399. https://www.proquest.com/scholarly-journals/functional-form-approach-calibration/docview/1266846662/se-2.

Isaki, C. T., and W. A. Fuller. 1982. “Survey Design Under the Regression Super-Population Model.” Journal of the American Statistical Association 77 (377): 89–96.

Kott, Phillip S., and Ted Chang. 2010. “Using Calibration Weighting to Adjust for Nonignorable Unit Nonresponse.” Journal of the American Statistical Association 105 (491): 1265–75. https://doi.org/10.1198/jasa.2010.tm09016.

Kott, Phillip S., and Dan Liao. 2017. “Calibration Weighting for Nonresponse That Is Not Missing at Random: Allowing More Calibration Than Response-Model Variables.” Journal of Survey Statistics and Methodology 5 (2): 159–74. https://doi.org/10.1093/jssam/smx003.

Author

Maciej Ostapiuk and Maciej Beręsewicz based on Kott and Chang (2010) and Kott and Liao (2017)

Examples

## Generate the data to formulate response model, calibration variables and target variable

set.seed(123)
n <- 10000
x1 <- rlnorm(n, 0, 1)
x2 <- rexp(n, 1)
x3 <-  rpois(n,2)
y <- 1 + x1 + x2 + rnorm(n)
pr <- plogis(1 + 0.5*x1 - 0.5*y)
pop_data <- data.frame(x1, x2, y, pr)
totals <- c(N = n, colSums(pop_data[, c("x1", "x2", "y")]))
flag <- rbinom(n, 1, pop_data$pr)
sample <- pop_data[flag == 1, ]
sample$d <- n/nrow(sample)

## case when dim(Xs) == dim(Zs)
g <- gencal(Xs = stats::model.matrix(~ x1 + x2, data = sample),
            Zs = stats::model.matrix(~ x1 + y, data = sample),
            d = sample$d,
            pop_totals = totals,
            method = "raking",
            eps = .Machine$double.eps,
            maxit = 100,
            tol = 1e-8)

head(g) #display only couple of first output values
#> [1] 1.1800080 2.6205677 1.0000360 1.1813942 0.6259329 0.7588780

## case when dim(Xs) > dim(Zs)
g1 <-gencal(Xs = stats::model.matrix(~ x1 + x2, data = sample),
            Zs = stats::model.matrix(~ y, data = sample),
            d = sample$d,
            pop_totals = totals,
            method = "raking",
            eps = .Machine$double.eps,
            maxit = 100,
            tol = 1e-8)
#>    [1] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>    [9] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>   [17] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>   [25] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>   [33] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>   [41] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>   [49] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>   [57] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>   [65] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>   [73] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>   [81] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>   [89] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>   [97] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [105] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [113] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [121] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [129] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [137] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [145] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [153] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [161] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [169] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [177] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [185] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [193] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [201] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [209] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [217] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [225] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [233] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [241] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [249] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [257] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [265] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [273] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [281] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [289] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [297] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [305] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [313] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [321] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [329] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [337] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [345] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [353] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [361] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [369] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [377] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [385] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [393] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [401] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [409] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [417] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [425] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [433] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [441] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [449] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [457] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [465] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [473] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [481] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [489] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [497] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [505] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [513] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [521] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [529] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [537] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [545] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [553] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [561] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [569] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [577] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [585] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [593] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [601] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [609] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [617] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [625] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [633] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [641] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [649] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [657] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [665] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [673] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [681] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [689] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [697] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [705] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [713] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [721] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [729] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [737] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [745] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [753] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [761] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [769] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [777] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [785] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [793] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [801] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [809] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [817] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [825] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [833] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [841] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [849] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [857] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [865] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [873] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [881] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [889] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [897] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [905] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [913] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [921] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [929] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [937] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [945] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [953] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [961] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [969] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [977] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [985] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#>  [993] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1001] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1009] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1017] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1025] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1033] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1041] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1049] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1057] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1065] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1073] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1081] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1089] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1097] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1105] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1113] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1121] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1129] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1137] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1145] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1153] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1161] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1169] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1177] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1185] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1193] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1201] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1209] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1217] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1225] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1233] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1241] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1249] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1257] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1265] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1273] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1281] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1289] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1297] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1305] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1313] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1321] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1329] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1337] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1345] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1353] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1361] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1369] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1377] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1385] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1393] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1401] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1409] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1417] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1425] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1433] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1441] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1449] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1457] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1465] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1473] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1481] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1489] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1497] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1505] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1513] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1521] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1529] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1537] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1545] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1553] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1561] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1569] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1577] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1585] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1593] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1601] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1609] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1617] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1625] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1633] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1641] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1649] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1657] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1665] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1673] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1681] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1689] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1697] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1705] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1713] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1721] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1729] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1737] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1745] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1753] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1761] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1769] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1777] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1785] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1793] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1801] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1809] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1817] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1825] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1833] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1841] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1849] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1857] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1865] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1873] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1881] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1889] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1897] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1905] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1913] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1921] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1929] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1937] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1945] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1953] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1961] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1969] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1977] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1985] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [1993] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2001] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2009] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2017] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2025] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2033] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2041] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2049] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2057] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2065] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2073] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2081] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2089] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2097] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2105] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2113] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2121] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2129] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2137] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2145] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2153] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2161] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2169] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2177] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2185] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2193] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2201] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2209] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2217] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2225] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2233] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2241] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2249] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2257] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2265] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2273] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2281] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2289] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2297] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2305] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2313] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2321] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2329] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2337] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2345] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2353] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2361] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2369] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2377] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2385] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2393] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2401] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2409] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2417] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2425] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2433] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2441] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2449] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2457] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2465] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2473] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2481] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2489] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2497] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2505] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2513] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2521] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2529] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2537] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2545] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2553] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2561] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2569] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2577] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2585] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2593] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2601] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2609] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2617] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2625] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2633] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2641] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2649] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2657] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2665] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2673] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2681] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2689] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2697] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2705] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2713] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2721] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2729] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2737] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2745] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2753] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2761] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2769] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2777] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2785] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2793] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2801] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2809] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2817] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2825] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2833] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2841] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2849] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2857] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2865] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2873] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2881] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2889] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2897] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2905] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2913] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2921] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2929] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2937] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2945] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2953] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2961] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2969] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2977] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2985] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [2993] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3001] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3009] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3017] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3025] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3033] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3041] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3049] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3057] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3065] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3073] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3081] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3089] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3097] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3105] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3113] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3121] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3129] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3137] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3145] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3153] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3161] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3169] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3177] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3185] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3193] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3201] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3209] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3217] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3225] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3233] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3241] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3249] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3257] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3265] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3273] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3281] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3289] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3297] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3305] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3313] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3321] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3329] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3337] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3345] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3353] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3361] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3369] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3377] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3385] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3393] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3401] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3409] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3417] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3425] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3433] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3441] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3449] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3457] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3465] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3473] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3481] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3489] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3497] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3505] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3513] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3521] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3529] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3537] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3545] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3553] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3561] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3569] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3577] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3585] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3593] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3601] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3609] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3617] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3625] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3633] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3641] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3649] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3657] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3665] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3673] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3681] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3689] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3697] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3705] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3713] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3721] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3729] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3737] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3745] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3753] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3761] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3769] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3777] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3785] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3793] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3801] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3809] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3817] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3825] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3833] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3841] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3849] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3857] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3865] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3873] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3881] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3889] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3897] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3905] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3913] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3921] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3929] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3937] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3945] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3953] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3961] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3969] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3977] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3985] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [3993] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4001] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4009] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4017] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4025] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4033] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4041] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4049] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4057] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4065] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4073] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4081] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4089] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4097] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4105] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4113] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4121] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4129] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4137] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4145] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4153] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4161] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4169] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4177] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4185] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4193] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4201] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4209] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4217] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4225] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4233] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4241] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4249] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4257] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4265] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4273] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4281] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4289] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4297] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4305] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4313] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4321] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4329] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4337] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4345] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4353] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4361] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4369] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4377] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4385] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4393] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4401] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4409] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4417] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4425] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4433] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4441] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4449] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4457] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4465] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4473] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4481] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4489] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4497] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4505] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4513] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4521] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4529] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4537] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4545] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4553] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4561] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4569] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4577] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4585] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4593] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4601] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4609] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4617] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4625] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4633] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4641] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4649] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4657] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4665] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4673] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4681] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4689] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4697] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4705] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4713] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4721] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4729] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4737] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4745] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4753] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4761] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4769] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4777] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4785] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4793] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4801] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4809] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4817] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4825] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4833] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4841] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4849] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4857] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4865] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4873] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4881] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4889] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4897] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4905] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4913] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4921] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4929] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4937] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4945] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4953] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4961] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4969] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4977] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4985] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [4993] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [5001] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [5009] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [5017] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [5025] 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309 1.985309
#> [5033] 1.985309 1.985309 1.985309 1.985309 1.985309
#> Error in gencal(Xs = stats::model.matrix(~x1 + x2, data = sample), Zs = stats::model.matrix(~y,     data = sample), d = sample$d, pop_totals = totals, method = "raking",     eps = .Machine$double.eps, maxit = 100, tol = 1e-08): object 'x_tilde' not found

head(g1) #display only couple of first output values
#> Error in eval(expr, envir, enclos): object 'g1' not found